% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroids.R
\docType{data}
\name{geom_centroids}
\alias{geom_centroids}
\alias{stat_centroids}
\alias{StatCentroids}
\title{Geographic centroid of administrative areas}
\usage{
geom_centroids(
  mapping = ggplot2::aes(),
  data = NULL,
  stat = "centroids",
  position = "circle_repel_sf",
  ...,
  feature_type = NA,
  inset = NULL,
  inset_copy = TRUE,
  inset_clip = TRUE,
  na.rm = TRUE,
  show.legend = "point",
  inherit.aes = TRUE
)

stat_centroids(
  mapping = NULL,
  data = NULL,
  geom = "sf",
  position = "circle_repel",
  ...,
  feature_type = NA,
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, stat, geom, position, na.rm, show.legend, inherit.aes, ...}{See \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}.}

\item{feature_type}{Type of map feature. See \code{\link[cartographer:feature_types]{feature_types()}} for a list of
registered types. If \code{NA}, the type is guessed based on the values in
\code{feature_names}.}

\item{inset}{Inset configuration; see \code{\link[ggmapinset:configure_inset]{configure_inset()}}.}

\item{inset_copy}{Draw both the base layer and the inset layer using the same
configuration. Only relevant when \code{inset} is specified.}

\item{inset_clip}{Clip away the part of the base layer corresponding to the
inset frame. Only relevant when \code{inset} is specified.}
}
\value{
A ggplot layer.
}
\description{
Assigns each point a longitude and latitude corresponding to the geographic
centre of its administrative area. This means that all points in the same
area will overlap. The default \code{position} uses \code{\link[=position_circle_repel]{position_circle_repel()}}
to repel the points outwards with an amount controllable with its
\code{density} parameter.
}
\section{Aesthetics}{

The \code{location} aesthetic is required.
\code{geom_centroids()} understands the same aesthetics as \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.
}

\section{Computed variables}{

\describe{
\item{cendroid_longitude}{longitude of the administrative region's centroid}
\item{centroid_latitude}{latitude of the administrative region's centroid}
}
}

\examples{
library(ggplot2)

cartographer::nc_type_example[1:49,] |>
  ggplot(aes(location = county)) +
  geom_boundaries(feature_type = "sf.nc") +
  geom_centroids(aes(colour = type, scale = 6), size = 0.5)
}
\keyword{datasets}
