% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_automap_zoom.R
\name{coord_automap_zoom}
\alias{coord_automap_zoom}
\title{Zoom a map to show only certain features}
\usage{
coord_automap_zoom(include = NULL, include_data = TRUE, feature_type = NA, ...)
}
\arguments{
\item{include}{Vector of feature names that should be shown on the map.}

\item{include_data}{Scalar logical, if true then all features with data are
also included.}

\item{feature_type}{Type of map feature. See \code{\link[cartographer:feature_types]{feature_types()}} for a list of
registered types. If \code{NA}, the type is guessed based on the values in
\code{feature_names}.}

\item{...}{Additional arguments passed to \code{\link[=coord_automap]{coord_automap()}}.}
}
\value{
A zoom specification that can be added to a ggplot object with \link[ggplot2:gg-add]{ggplot2::\%+\%}.
}
\description{
This is a wrapper around \code{\link[=coord_automap]{coord_automap()}} that automatically calculates
coordinate limits based on the data and/or any additional locations. The
bounding box will be calculated to encompass all of the \code{include}d
locations.
}
\details{
This should be added to the plot \emph{after} the call to one of the ggautomap
geoms. It will copy the \code{location} aethetic mapping from the first such
layer in the plot. If there is no such layer, it will attempt to use the data
and \code{location} mapping found at the top level \code{ggplot()} call.
}
\examples{
library(ggplot2)

# zoom in on locations that have data:
cartographer::nc_type_example_2 |>
  ggplot(aes(location = county)) +
  geom_boundaries(feature_type = "sf.nc") +
  geom_choropleth() +
  coord_automap_zoom()

# or just zoom in on specific locations regardless of the data:
cartographer::nc_type_example_2 |>
  ggplot(aes(location = county)) +
  geom_boundaries(feature_type = "sf.nc") +
  coord_automap_zoom(include = c("Rowan", "Polk"), include_data = FALSE)
}
\seealso{
\code{\link[=coord_automap]{coord_automap()}}
}
