% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_sf_zoom.R
\name{coord_sf_zoom}
\alias{coord_sf_zoom}
\title{Zoom a map to show only certain features.}
\usage{
coord_sf_zoom(include = NULL, include_data = TRUE, ...)
}
\arguments{
\item{include}{Vector of feature names that should be shown on the map.}

\item{include_data}{Scalar logical, if true then all features with data are
also included.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:ggsf]{ggplot2::coord_sf()}}.}
}
\description{
This is a wrapper around \code{\link[ggplot2:ggsf]{ggplot2::coord_sf()}} that automatically calculates
coordinate limits based on the data and/or any additional locations. The
bounding box will be calculated to encompass all of the \code{include}d
locations.
}
\details{
This should be added to the plot \emph{after} the call to one of the ggautomap
geoms. It will copy the data and \code{feature_type} from the first such
layer in the plot. If there is no such layer, it will attempt to guess the
feature type and use the data and \code{location} mapping found at the top
level \code{ggplot()} call.
}
