% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggautomap-global.R
\name{feature_names}
\alias{feature_names}
\title{List known feature names}
\usage{
feature_names(feature_type)
}
\arguments{
\item{feature_type}{Type of map feature. See \link{feature_types} for a list of
registered types.}
}
\value{
Character vector of names.
}
\description{
List known feature names
}
