% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggautomap-global.R
\name{feature_types}
\alias{feature_types}
\title{List known feature types}
\usage{
feature_types()
}
\description{
Each feature type corresponds to map data that has been registered.
}
\seealso{
register_map
}
