% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.R
\docType{data}
\name{geom_pie}
\alias{geom_pie}
\alias{GeomPie}
\alias{StatCentroidPie}
\title{Summarise regions with pie charts}
\usage{
geom_pie(
  mapping = ggplot2::aes(),
  data = NULL,
  stat = "centroid_pie",
  position = "identity",
  ...,
  feature_type = NA,
  pie_radius = 1,
  proportional = FALSE,
  na.rm = TRUE,
  show.legend = TRUE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{See \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}.}

\item{feature_type}{Type of map feature. See \link{feature_types} for a list of
registered types. If \code{NA}, the type is guessed based on the values in
the data.}

\item{pie_radius}{Scale the side of all pies}

\item{proportional}{If \code{TRUE}, scale the pies by the number of rows in
each region. The radius of each pi is proportional to the count.}
}
\description{
Summarise regions with pie charts
}
\keyword{datasets}
