% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{into_sf}
\alias{into_sf}
\title{Convert input data frame into a spatial data frame}
\usage{
into_sf(x, location, feature_type = NA, geom_name = "geometry")
}
\arguments{
\item{x}{Data frame with a location column.}

\item{location}{Name of the location column (tidy evaluation).}

\item{feature_type}{Feature type of the location column. If NA (the default),
the type is guessed from the values in the location column.}

\item{geom_name}{Name for the new column to contain the geometry.}
}
\value{
A spatial data frame containing all of the columns from the input
data frame.
}
\description{
Convert input data frame into a spatial data frame
}
\examples{
data(nc_type_example)

into_sf(nc_type_example, location)
}
