% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc_type_example}
\alias{nc_type_example}
\title{Example dataset for ggautomap illustration}
\format{
A data frame with 50 rows and 2 columns:
\describe{
\item{location}{Location names that match the NAME field of the nc dataset}
\item{type}{Arbitrary field to map to colour/fill aesthetics}
}
}
\usage{
nc_type_example
}
\description{
This dataset contains random data compatible with the \code{sf.nc} example
map data for illustrating ggautomap's features.
}
\keyword{datasets}
