% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_circle_repel.R
\docType{data}
\name{position_circle_repel}
\alias{position_circle_repel}
\alias{position_circle_repel_sf}
\alias{PositionCircleRepel}
\alias{PositionCircleRepelSf}
\title{Pack overlapping points into a circle}
\usage{
position_circle_repel(scale = 1/4)

position_circle_repel_sf(scale = 10)
}
\arguments{
\item{scale}{Scale of packing around the central point. Note that you can
alternatively set this to a data column by using the \code{scale} aesthetic.
This is in data units, so for the \code{_sf} variant it will depend on the
units specified by the coordinate reference system.}
}
\description{
This position looks for any points with identical \code{x} and \code{y}
positions and packs them in a circle around the original point. The \code{_sf}
version applies the position adjustment in projected coordinates.
}
\details{
Note that extreme choices of \code{scale} may cause errors.
}
\keyword{datasets}
