% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggautomap-global.R
\name{register_map}
\alias{register_map}
\title{Register a new feature type}
\usage{
register_map(
  feature_type,
  data,
  feature_column,
  aliases = NULL,
  outline = NULL
)
}
\arguments{
\item{feature_type}{Name of the type. Should include the package name to
avoid clashes if registered in a package.}

\item{data}{A simple feature data frame with the map data.}

\item{feature_column}{Name of the column of \code{data} that contains the
feature names.}

\item{aliases}{Optional named character vector or list that maps aliases to
values that appear in the feature column. This allows abbreviations or
alternative names to be supported.}

\item{outline}{Optional sf geometry containing just the outline of the map.}
}
\description{
This adds a new feature type that can then be used by all the geoms in this
package. If registering from another package, this should occur in the
\code{.onLoad()} hook in the package.
}
