% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggautomap-global.R
\name{resolve_feature_type}
\alias{resolve_feature_type}
\title{Guess the feature type if it was missing}
\usage{
resolve_feature_type(feature_type, locations, context)
}
\arguments{
\item{feature_type}{Type of map feature. See \link{feature_types} for a list of
registered types. If \code{NA}, the type is guessed based on the values in
the data.}

\item{locations}{Character vector of location names in the data.}

\item{context}{Name of the calling function, for inclusion in error message
if it's not possible to unambiguously guess the feature type.}
}
\description{
Guess the feature type if it was missing
}
