% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{geom_boundaries}
\alias{geom_boundaries}
\title{Map state boundaries}
\usage{
geom_boundaries(
  mapping = ggplot2::aes(),
  data = NULL,
  stat = "sf",
  position = "identity",
  ...,
  feature_type = NULL,
  inset = NULL,
  inset_copy = TRUE,
  na.rm = FALSE,
  outline_colour = "#666666",
  outline_linewidth = NULL,
  show.legend = NA,
  inherit.aes = FALSE
)
}
\arguments{
\item{mapping, stat, position, na.rm, show.legend, inherit.aes, ...}{See \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}.}

\item{data}{Ignored (this geometry always uses the registered geographic data).}

\item{feature_type}{Type of map feature. See \code{\link[=feature_types]{feature_types()}} for a list of
registered types. If \code{NA}, the type is guessed based on the values in
the data.}

\item{inset}{Inset configuration; see \code{\link[ggmapinset:configure_inset]{configure_inset()}}.}

\item{inset_copy}{Draw both the base layers and the inset layers using the same
configuration. Only relevant when \code{inset} is specified.}

\item{outline_colour, outline_linewidth}{Override the aesthetics of the state outline.
If \code{NULL}, the aesthetic is inherited from \code{mapping}.}
}
\description{
Small wrappers around \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}} that load the appropriate
geometry dataset and add inset support.
}
\details{
As well as the chosen feature boundaries, the outline of the map is
drawn separately, with the possibility to override its \code{colour} and
\code{linewidth} aesthetics.
}
\examples{
library(ggplot2)

ggplot(nc_type_example, aes(location = location)) +
geom_boundaries(feature_type = "sf.nc")

}
