% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\docType{data}
\name{geom_choropleth}
\alias{geom_choropleth}
\alias{stat_choropleth}
\alias{StatChoropleth}
\title{Associate regions with counts}
\usage{
geom_choropleth(
  mapping = ggplot2::aes(),
  data = NULL,
  stat = "choropleth",
  position = "identity",
  ...,
  feature_type = NA,
  inset = NULL,
  inset_copy = TRUE,
  inset_clip = FALSE,
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_choropleth(
  mapping = NULL,
  data = NULL,
  geom = "sf",
  position = "identity",
  ...,
  feature_type = NA,
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, stat, geom, position, na.rm, show.legend, inherit.aes, ...}{See \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}.}

\item{feature_type}{Type of map feature. See \code{\link[cartographer:feature_types]{feature_types()}} for a list of
registered types. If \code{NA}, the type is guessed based on the values in
\code{feature_names}.}

\item{inset}{Inset configuration; see \code{\link[ggmapinset:configure_inset]{configure_inset()}}.}

\item{inset_copy}{Draw both the base layer and the inset layer using the same
configuration. Only relevant when \code{inset} is specified.}

\item{inset_clip}{Clip away the part of the base layer corresponding to the
inset frame. Only relevant when \code{inset} is specified.}
}
\value{
A ggplot layer.
}
\description{
Counts the number of occurrences of each location, then by default maps the
count to the fill aesthetic. If your data has only one row per location and
some other field that you'd like to map to aesthetics, use
\code{\link[ggplot2:ggsf]{geom_sf()}} or \code{\link[ggmapinset:geom_sf_inset]{geom_sf_inset()}}
with \code{stat = "automap"} instead.
}
\details{
Note that choropleths have a tendency to be misleading by emphasising
geographically larger areas.
}
\section{Aesthetics}{

The \code{location} aesthetic is required.
\code{geom_choropleth()} understands the same aesthetics as \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}.
}

\section{Computed variables}{

\describe{
\item{count}{rows matching the region}
\item{geometry}{\code{sf} geometry column}
\item{...}{limits as computed by \code{\link[ggplot2:ggsf]{ggplot2::stat_sf()}}}
}
}

\examples{
library(ggplot2)

cartographer::nc_type_example[1:49,] |>
  ggplot(aes(location = county)) +
  geom_boundaries(feature_type = "sf.nc") +
  geom_choropleth() +
  scale_fill_steps(low = "#e6f9ff", high = "#00394d", na.value = "white")
}
\keyword{datasets}
