% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\docType{data}
\name{geom_choropleth}
\alias{geom_choropleth}
\alias{stat_choropleth}
\alias{StatChoropleth}
\title{Associate regions with counts}
\usage{
geom_choropleth(
  mapping = ggplot2::aes(),
  data = NULL,
  stat = "choropleth",
  position = "identity",
  ...,
  feature_type = NA,
  inset = NULL,
  inset_copy = TRUE,
  inset_clip = FALSE,
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_choropleth(
  mapping = NULL,
  data = NULL,
  geom = "sf",
  position = "identity",
  ...,
  feature_type = NA,
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, stat, geom, position, na.rm, show.legend, inherit.aes, ...}{See \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}.}

\item{feature_type}{Type of map feature. See \code{\link[=feature_types]{feature_types()}} for a list of
registered types. If \code{NA}, the type is guessed based on the values in
the data.}

\item{inset}{Inset configuration; see \code{\link[ggmapinset:configure_inset]{configure_inset()}}.}

\item{inset_copy}{Draw both the base layers and the inset layers using the same
configuration. Only relevant when \code{inset} is specified.}

\item{inset_clip}{When an inset is drawn, place included points only in the inset.}
}
\description{
Instead of representing each row separately, aggregate them as counts that
can be used to draw choropleths. Note that choropleths have a tendency to be
misleading by emphasising geographically larger areas.
}
\section{Aesthetics}{

The \code{location} aesthetic is required.
\code{geom_choropleth()} understands the same aesthetics as \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}.
}

\section{Computed variables}{

\describe{
\item{count}{rows matching the region}
\item{geometry}{\code{sf} geometry column}
\item{...}{limits as computed by \code{\link[ggplot2:ggsf]{ggplot2::stat_sf()}}}
}
}

\examples{
library(ggplot2)

ggplot(nc_type_example, aes(location = location)) +
geom_boundaries(feature_type = "sf.nc") +
geom_choropleth() +
scale_fill_steps(low = "#e6f9ff", high = "#00394d", na.value = "white")

}
\keyword{datasets}
