% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggautomap-global.R
\name{feature_names}
\alias{feature_names}
\title{List known feature names}
\usage{
feature_names(feature_type)
}
\arguments{
\item{feature_type}{Type of map feature. See \link{feature_types} for a list of
registered types.}
}
\value{
Character vector of names.
}
\description{
This gives the list of feature names that are part of the specified map data.
When matching your data column to the map data, the names are resolved by
checking for the first match using:
\enumerate{
\item case sensitive match, then
\item case sensitive match using aliases, then
\item case insensitive match, then
\item case insensitive match using aliases.
}
}
