% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggautomap-package.R
\docType{package}
\name{ggautomap-package}
\alias{ggautomap}
\alias{ggautomap-package}
\title{Create Maps From a Column of Place Names}
\description{
Mapping geometries based on ggplot2 for easy maps.
This package provides ggplot geoms that make maps from a data frame where
one of the columns contains place names. The map data must be registered
either by the user or by another package (e.g. \code{{nswgeo}}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cidm-ph/ggautomap}
  \item \url{https://cidm-ph.github.io/ggautomap/}
  \item Report bugs at \url{https://github.com/cidm-ph/ggautomap/issues}
}

}
\author{
\strong{Maintainer}: Carl Suster \email{Carl.Suster@health.nsw.gov.au} (\href{https://orcid.org/0000-0001-7021-9380}{ORCID})

Other contributors:
\itemize{
  \item Thomas Lin Pedersen (Original author of some code in pie.R taken from 'ggforce') [contributor]
  \item Western Sydney Local Health District, NSW Health [copyright holder]
}

}
