% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_automap_coords.R
\name{stat_automap_coords}
\alias{stat_automap_coords}
\title{Attach spatial data with cartographer and reduce to coordinates}
\usage{
stat_automap_coords(
  mapping = NULL,
  data = NULL,
  geom = "sf",
  position = "identity",
  ...,
  feature_type = NA,
  na.rm = TRUE,
  fun.geometry = NULL,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, geom, position, na.rm, show.legend, inherit.aes, ...}{See \code{\link[ggplot2:stat_sf_coordinates]{ggplot2::stat_sf_coordinates()}}.}

\item{feature_type}{Type of map feature. See \code{\link[cartographer:feature_types]{feature_types()}} for a list of
registered types. If \code{NA}, the type is guessed based on the values in
\code{feature_names}.}

\item{fun.geometry}{A function that takes a \code{sfc} object and returns a \code{sfc_POINT} with the
same length as the input. If \code{NULL}, \code{function(x) sf::st_point_on_surface(sf::st_zm(x))}
will be used. Note that the function may warn about the incorrectness of
the result if the data is not projected, but you can ignore this except
when you really care about the exact locations.}
}
\value{
A plot layer
}
\description{
Use \code{\link{cartographer}} to attach a spatial column to the data based
on place names in another column. The spatial data is then reduced to
coordinates in the same way as \code{\link[ggplot2:stat_sf_coordinates]{stat_sf_coordinates()}}.
The result can then be used by \code{\link[ggplot2:ggsf]{geom_sf()}} or
\code{\link[ggmapinset:geom_sf_inset]{geom_sf_inset()}} or any geom that needs \code{x} and
\code{y} aesthetics.
}
\section{Computed variables}{

\describe{
\item{x}{X dimension of the simple feature}
\item{y}{Y dimension of the simple feature}
}
}

\examples{
library(ggplot2)

events <- data.frame(
  county = c("Mecklenburg", "Carteret", "Moore", "Caldwell"),
  proportion_A = c(0.1, 0.8, 0.0, 0.6))

ggplot(events, aes(location = county)) +
  geom_sf(aes(fill = proportion_A), stat = "automap") +
  geom_label(aes(label = county), stat = "automap_coords")
}
