% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_automap_coords.R
\name{stat_automap_coords}
\alias{stat_automap_coords}
\title{Attach coordinates with 'cartographer'}
\usage{
stat_automap_coords(
  mapping = NULL,
  data = NULL,
  geom = "sf",
  position = "identity",
  ...,
  feature_type = NA,
  na.rm = TRUE,
  fun.geometry = NULL,
  crs_working = NULL,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, geom, position, na.rm, show.legend, inherit.aes, ...}{See \code{\link[ggplot2:stat_sf_coordinates]{ggplot2::stat_sf_coordinates()}}.}

\item{feature_type}{Type of map feature. See \code{\link[cartographer:feature_types]{feature_types()}} for a list of
registered types. If \code{NA}, the type is guessed based on the values in
\code{feature_names}.}

\item{fun.geometry}{Function to transform the geometry into one point per row.
See \code{\link[ggplot2:stat_sf_coordinates]{stat_sf_coordinates()}}; the difference
is that the spatial data are first transformed to \code{crs_working} to
reduce the risk of projection-related distortions.}

\item{crs_working}{Coordinate reference system where \code{fun.geometry} is
applied. If not provided, \code{\link[=crs_eqc]{crs_eqc()}} is used with a midpoint based on the
map data.}
}
\value{
A plot layer
}
\description{
Use \code{\link{cartographer}} to attach a spatial column to the data based
on place names in another column. The spatial data is then reduced to
coordinates in the same way as \code{\link[ggplot2:stat_sf_coordinates]{stat_sf_coordinates()}}.
The result can then be used by \code{\link[ggplot2:ggsf]{geom_sf()}} or
\code{\link[ggmapinset:geom_sf_inset]{geom_sf_inset()}} or any geom that needs \code{x} and
\code{y} aesthetics.
}
\section{Computed variables}{

\describe{
\item{geometry}{\code{sf} geometry column containing the points}
\item{x}{X dimension of the simple feature}
\item{y}{Y dimension of the simple feature}
}
}

\examples{
library(ggplot2)

events <- data.frame(
  county = c("Mecklenburg", "Carteret", "Moore", "Caldwell"),
  proportion_A = c(0.1, 0.8, 0.0, 0.6))

ggplot(events, aes(location = county)) +
  geom_sf(aes(fill = proportion_A), stat = "automap") +
  geom_label(aes(label = county), stat = "automap_coords")
}
